refactor: modularize MCP server, add resource groups, improve examples docs
Split the monolithic 1950-line mcp/run_mcp.py into a modular package:
- mcp/server.py, audit.py, iam.py, cli_runner.py, version.py
- mcp/tools/ (jobs, capacity, inference, costs, stacks, storage, models)
- mcp/resources/ (docs, source, k8s, iam, infra, ci, demos, clients, scripts, tests, config)

New MCP resource groups:
- tests:// for test suite docs, infrastructure helpers, and BATS tests
- config:// for cdk.json, parsed feature toggles, and env var docs
- docs://gco/examples/guide for manifest creation patterns and metadata

Enhanced example indexing:
- EXAMPLE_METADATA dict with category, summary, GPU, opt-in, and submission
  command for all 28 examples
- docs://gco/examples/{name} returns a metadata header before the YAML
- docs://gco/index categorizes examples by type

Examples README rewritten:
- Table of contents and quick reference table covering all 28 examples
- All examples and tables sorted alphabetically
- Cross-links to relevant docs (INFERENCE, SCHEDULERS, CUSTOMIZATION)

Updated EFA bandwidth stats:
- P4d 400 Gbps, P5/Trn2 3,200 Gbps, P6-B200 3.2 Tbps, P6e-GB200 28.8 Tbps
- Added p5e and p6e-gb200 rows to docs/CUSTOMIZATION.md
- Corrected Trn1 from 800 Gbps to 1,600 Gbps

Tests: 199 passing (168 existing updated + 31 new)
