+++ mcp/tools/storage.py
@@ -1,0 +1,34 @@
+"""File storage MCP tools."""
+
+from audit import audit_logged
+import cli_runner
+from server import mcp
+
+
+@mcp.tool()
+@audit_logged
+def list_storage_contents(region: str, path: str = "/") -> str:
+    """List contents of shared EFS storage.
+
+    Args:
+        region: AWS region.
+        path: Directory path to list (default: root).
+    """
+    args = ["files", "ls", "-r", region]
+    if path != "/":
+        args.append(path)
+    return cli_runner._run_cli(*args)
+
+
+@mcp.tool()
+@audit_logged
+def list_file_systems(region: str | None = None) -> str:
+    """List EFS and FSx file systems.
+
+    Args:
+        region: Specific region, or omit for all.
+    """
+    args = ["files", "list"]
+    if region:
+        args += ["-r", region]
+    return cli_runner._run_cli(*args)
